refactor: move activation checkpointing wrapper to blocks
